requirements.txt
certifi==2018.10.15
chardet==3.0.4
clearbit==0.1.7
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.3.2
pyhunter==0.3
PyJWT==1.6.4
pytz==2018.5
requests==2.20.0
uritemplate==3.0.0
urllib3==1.24
